corechecks/network: enable Go network check on AIX
gopsutil now implements ProtoCounters on AIX (via netstat -s), IOCounters
(via netstat -i), and Connections (via netstat -anv), providing everything
the Go network check needs.

Changes:
- Extend network.go build tag from linux to linux || aix
- Extract netstatTCPExtCounters() to netstat_tcpext_linux.go (Linux only,
  reads /proc/net/netstat which does not exist on AIX)
- Add netstat_tcpext_aix.go returning nil, nil (no-op; TCPExt counters
  are silently skipped when nil is returned by the check)
- Bypass the network_check.use_core_loader gate on AIX — there is no
  Python network check on AIX, so the Go check is unconditionally enabled
- Add "network" to AIX_CORECHECKS in tasks/core_checks.py and
  cmd/agent/dist/core_checks.bzl so it is bundled in the AIX agent binary

The protocolsMetricsMapping is shared unchanged; keys absent from AIX's
netstat -s output (e.g. ListenOverflows, TCPBacklogDrop) are silently
skipped by submitProtocolMetrics.
