+++ pkg/collector/corechecks/net/network/netstat_tcpext_aix.go
@@ -1,0 +1,12 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build aix
+
+package network
+
+func netstatTCPExtCounters() (map[string]int64, error) {
+	return nil, nil
+}
